Fix codestyle and some coding issues
